POST /threeStyleTable を別ファイルに切り出した
